delete right only for admin

+++ Views/clients.view.php
@@ -2,8 +2,8 @@
 
 use Clive\Core\Mantle\Paginator;
 
- include_once 'base.view.php'; 
- include_once 'sections/dash-nav.view.php';?>
+include_once 'base.view.php';
+include_once 'sections/dash-nav.view.php'; ?>
 
 
 
@@ -41,44 +41,47 @@ use Clive\Core\Mantle\Paginator;
                                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M2.458 12C3.732 7.943 7.523 5 12 5c4.478 0 8.268 2.943 9.542 7-1.274 4.057-5.064 7-9.542 7-4.477 0-8.268-2.943-9.542-7z" />
                                         </svg>
                                     </a>
+
                                     <a href="edituser?user_id<?= "$user->user_id&uname=$user->username" ?>">
                                         <svg xmlns="http://www.w3.org/2000/svg" class="hover:text-green-600 text-green-400 w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z" />
                                         </svg>
                                     </a>
-                                    <a href="deleteuser?user_id<?= "$user->user_id&uname=$user->username" ?>">
-                                        <svg xmlns="http://www.w3.org/2000/svg" class="hover:text-red-600 text-red-400 h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                                        </svg>
-                                    </a>
+                                    <?php if (isAdmin()) : ?>
+                                        <a href="deleteuser?user_id<?= "$user->user_id&uname=$user->username" ?>">
+                                            <svg xmlns="http://www.w3.org/2000/svg" class="hover:text-red-600 text-red-400 h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                                                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
+                                            </svg>
+                                        </a>
+                                    <?php endif; ?>
                                 </p>
                             </td>
                         </tr>
                     <?php endforeach; ?>
-                 
+
                 </tbody>
             </table>
             <div class="border border-t bg-white px-4 py-3 flex items-center justify-between sm:px-6">
-              
-              <div class="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
-                <div>
-                  <p class="text-sm leading-5 text-gray-700">
-                    Showing
-                    <span class="font-medium" ><?= Paginator::$start;?></span>
-                    to
-                    <span class="font-medium"><?= Paginator::$end;?></span>
-                    of
-                    <span class="font-medium" ><?= count($allusers)?></span>
-                    results
-                  </p>
+
+                <div class="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
+                    <div>
+                        <p class="text-sm leading-5 text-gray-700">
+                            Showing
+                            <span class="font-medium"><?= Paginator::$start; ?></span>
+                            to
+                            <span class="font-medium"><?= Paginator::$end; ?></span>
+                            of
+                            <span class="font-medium"><?= count($allusers) ?></span>
+                            results
+                        </p>
+                    </div>
+                    <div>
+                        <span class="relative z-0 inline-flex shadow-sm">
+                            <?php Paginator::showLinks($allusers); ?>
+                        </span>
+                    </div>
                 </div>
-                <div>
-                  <span class="relative z-0 inline-flex shadow-sm">
-                  <?php Paginator::showLinks($allusers);?>
-                  </span>
-                </div>
-              </div>
-            </div>          
+            </div>
         </div>
     </div>
 </div>
